refactor : voice 패키지 정리
