+++ src/main/java/inu/codin/codin/domain/board/voice/repository/VoiceRepository.java
@@ -1,7 +1,7 @@
-package inu.codin.codin.domain.voice.repository;
+package inu.codin.codin.domain.board.voice.repository;
 
 import inu.codin.codin.common.dto.Department;
-import inu.codin.codin.domain.voice.entity.VoiceEntity;
+import inu.codin.codin.domain.board.voice.entity.VoiceEntity;
 import org.bson.types.ObjectId;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
